feat(platform): T11 Settings UI 재구성 (Jira/GitHub sub-tab + connect 폼 분리)
- src/sidepanel/tabs/connect/JiraConnectForm.tsx (신규):
  - 기존 SettingsTab Jira UI를 외과적 추출 — behaviour 동일
  - PageShell 래퍼는 부모로 이전, 본 컴포넌트는 PageScroll/PageFooter만
- src/sidepanel/tabs/connect/GithubConnectForm.tsx (신규):
  - 빈 상태: GitHub로 로그인 버튼(OAuth) + Personal Access Token 섹션(Input + Save)
  - github.oauth.available 메시지로 OAuth 가능 여부 동적 체크 — env 미설정 시 버튼 disabled + 안내
  - github.startOAuth → GithubOAuthAuth → setAccount("github", {...})
  - github.testPat → viewer 검증 → setAccount("github", { auth: { kind: "pat", ... } })
  - DISMISS_PATTERNS로 사용자 취소는 에러 표시 안 함
  - 연결됨 상태: viewer 카드(login + OAuth/PAT 배지) + AlertDialog 기반 Disconnect
- src/sidepanel/tabs/SettingsTab.tsx (재작성): PageShell + shadcn Tabs[jira|github] thin wrapper. 본문은 connect 폼이 처리.
- pnpm typecheck/test 199 통과
